Refactor CollectionSchema.define for improved clarity
CollectionSchema.define(...) used to return a callback which then had to
be invoked with some mappings as an argument in order to generate the
schema. The callback was necessary for TyepScript's type inference to
work correctly.

I've now refactored this into a "fluent" style API. `define` now returns
an object with two methods: one for defining the indices and one for
definign a schema with no indices.