Import gRPC definitions from "file descriptor set"
Previously we imported definitions from *.proto files. The files were
required at runtime by grpc-js.

We hit a snag when deploying the Console application to Vercel. Vercel
creates JS bundles from your app - and only files that are directly
imported statically are available at runtime which meant that the
*.proto files could not be found and the app could not boot.

This commit defines a build step that uses `protoc` to create a "file
descriptor set" (binary) that contains the definitions for the entire
API as one single file.  It then builds a TypeScript file that exports a
literal Buffer containing the binary gRPC API definition.

grpc-js knows how to load API definitions from a file descriptor set
which is great 👍

This is a better solution than embedding every single *.protoc file
using a similar mechanism - with the individual files they *must* be
imported and loaded into grpc-js in the correct dependency order. There
is no need to be concerned about this when the entire set of files has
been combined into a single binary blob for us.

I decided to leave the *.proto files in the `dist` directory so that
when stashjs-grpc module is published to NPM the sources to the binary
blob remain available.