Bump stashjs version
Make console depend on the new version.

Made a change to publish-npm.sh so that it will only try to publish
public packages - previously it would try to publish stash-cli, which is
private, and the script would fail because we aren't paying NPM for the
ability to publish private packages.