Teach stash-cli how to grab workspace details from Console
This commit also introduces a new command `init`. Init has one mandatory
argument: the workspace ID.

It signs in via Auth0 device code authentication then hits the Console
API to grab the workspace details (in particular the cmk & naming key).

It then saves the config in ~/.cipherstash/<workspace-id>

There is more refactoring in a future commit but getting this landed
will unblock the grpc-router and end to end tests.