Add helper to cache expensive objects that depend on auth creds
Building a CipherSuite is an expensive operation - it requires talking
to KMS. So we want to avoid building them unnecessarily.

To this end, I have added a new function `withFreshCredentials` that
given an `AuthStrategy` and a callback that can build an object with
fresh credentials, it will return a `Memo` that will cache the built
value and rebuild it when ever it's asked for its object and the
credentials have since expired (they will be refreshed, and a new object
built).