Add errors module
This module defines all of the possible errors that can be produced by
StashJS.

All errors conform to the type StashJSError. Conforming types are
notionally sum types: for example, an AnalysisFailure's `cause` can be
an EncryptionFailure or an AuthenticationFailure.

The fact that they are sum types means that we get TypeScript support
for exhaustiveness checking when we need it, to ensure that all errors
are handled.

Every error type also has a corresponding constructor function with the
same name.