Add StashCLI create-collection command
This includes a parser & type checker for the JSON schema provided on
the command line.

The schema parser and type checker:
  - handles arbitrarily nested user-provided schemas
  - "typechecks" the index definitions
    - all referenced fields must exist on the user-provided type
    - the types of indexed fields must be appropriate for the kind of
      index

NOTE: there is a problem with using JTD to describe our schema for
describing user-provided types. JTD cannot express recursive types when
the keys are unknown.

Future options include:

- switching to JSON Schema to describe our schema language
- allow the user to use full-blown JTD to describe their schema (which
  will mean adding support for all JTD types to our indexes).