Fix: StashJS querying
An error was introcuced during error handling refactoring that returned
a promise in a QueryReply field instead of a resolved value.

This went unnoticed by the compiler because of a use of `any` in the
stashjs-examples query-collection code. (fix coming in another PR).

It also was not covered by an E2E test (not addressed).