Fix: MaxListenersExceededWarning: Possible EventEmitter memory leak detected
This error is sometimes seen on large imports.

The StreamWriter class used NodeJS's EventEmitter API to detect "error"
and "drain" events on streams. In particular it uses the "once" method
to handle an event one time which then removes the listener for you once
it has been fired.

The problem was that we are listening for not one but *two* events and
they were mutually exclusive - so every time a stream write was buffered
it would wait for one of the two events, one of which by definition
would never happen so the listeners for one of the events (the error
event in particular) were never freed.