Fix: collection schema parsing
TL;DR: Parse, don't validate!  (see:
https://lexi-lambda.github.io/blog/2019/11/05/parse-don-t-validate/)

Previously the parsing & type checking were seperate steps - and that
led to a bug in the create-collection command where the schema was
passed to the schema type checker as unvalidated raw user input.

Now, we have a function which accepts a string, parses into JSON and
type checks it in one step removing any possibility of forgetting to
validate.