Teach StashJS to handle connection resets
An ECONNRESET happens when user-level TCP code tries to use a previously
opened TCP connection that has been inactive for a period of time and
thus the operating system kernel has reaped it.

This happens on the Console app because the connection is initialised
when StashJS is initialised and the app code keeps hold of the
connection indefinitely.

Neither grpcjs nor NodeJS's built-in http2 library offer any support for
*transparently* reconnecting. (grpcjs's docs on this are confusing - it
claims to support "autoreconnect" but perusing the code leads me to
believe that's not a thing it supports, so maybe they have another
interpretation).

Note that auto reconnection is not provided for the `putStream` call
yet. It has a slightly different handler signature in the API middleware
and I haven't gotten around to this yet.

So the implementation I've written here adds another middleware to the
stash-api-async-result-wrapper code that checks if an error was a
connection reset and then immediately retries the request.

I've not tested this yet as it's hard to repro locally.