StashJS benchmarking
Implement benchmark for testing import speed.

This PR is designed to be equivalent the ActiveStash benchmark as much
as is feasible.

The schema defined here is supposed to the be the same scheme generated
by the ActiveStash DSL. The fields and the indexes should be identical.

Both this benchmark and the ActiveStash benchmark use randomly generated
data for the benchmark, using different frameworks so there will be some
differences.