Handle decryption errors when listing collections
This change prevents listing collections from returning errors by
filtering out collections that fail to decrypt. If CS_DEBUG=yes a
warning will be logged that the collection was filtered out. Otherwise,
the collection is silently dropped from results.

This is a short-term fix for incompatibilities in ciphertext formats
between StashJS and StashRB. Filtering out incompatible collections
means that StashJS will only list collections that it can work with.

The intent here is to implement a reasonable short-term solution without
changing other behaviour in StashJS. Note that this change still won't
allow for deleting collections created from StashRB, but StashJS should
be able to list collections that it created and clean up after itself
now (even if there are collections created by StashRB in the same
workspace).