Dockerise the build
Why did I do this?

1. Our builds had stopped working on a Mac because of use of `sed` in a
   build script - the Mac and Linux versions are different
2. asdf tool proliferation has become a thing across our projects
3. I can no longer get asdf postgres to install successfully on my mac

The base image (ubuntu:latest) will install a version appropriate to
your host architecture - i.e. x86_64 or arm64 so nobody has to pay an
unfair performance penalty.

It takes about 2 minutes to build the image. Once built, commands can be
executed in the container with the host repo mounted inside of it. There
does not appear to be any noticable performance drop in build times but
I have not run benchmarks.